Modify, Fix, fix grpc plugin gen code bug Modify, Test, add grpc plugin test

+++ example/grpc_common/python_example_proto_code/example_proto_by_option/user/user_pait_route.py
@@ -48,12 +48,18 @@ async def async_logout_user_route(
     token: str = Header.i(description="User Token"),
     req_id: str = Header.i(alias="X-Request-Id", default_factory=lambda: str(uuid4())),
 ) -> Any:
-    gateway = pait_context.get().app_helper.get_attributes("gateway_attr_gateway")
-    stub: user_pb2_grpc.UserStub = pait_context.get().app_helper.get_attributes("gateway_attr_User")
+    gateway = pait_context.get().app_helper.get_attributes("gateway_attr_user_by_option_gateway")
     request_dict: dict = request_pydantic_model.dict()
     request_dict["token"] = token
-    request_msg: user_pb2.LogoutUserRequest = gateway.get_msg_from_dict(user_p2p.LogoutUserRequest, request_dict)
-    grpc_msg: Empty = await stub.logout_user(request_msg, metadata=[("req_id", req_id)])
+    request_msg: user_pb2.LogoutUserRequest = gateway.get_msg_from_dict(user_pb2.LogoutUserRequest, request_dict)
+    loop: asyncio.AbstractEventLoop = asyncio.get_event_loop()
+    if loop != gateway.User_stub.logout_user._loop:  # type: ignore
+        raise RuntimeError(
+            "Loop is not same, "
+            "the grpc channel must be initialized after the event loop"
+            "of the api server is initialized"
+        )
+    grpc_msg: Empty = await gateway.User_stub.logout_user(request_msg, metadata=[("req_id", req_id)])
     return gateway.make_response(gateway.msg_to_dict(grpc_msg))
 
 
@@ -62,66 +68,61 @@ def logout_user_route(
     token: str = Header.i(description="User Token"),
     req_id: str = Header.i(alias="X-Request-Id", default_factory=lambda: str(uuid4())),
 ) -> Any:
-    gateway = pait_context.get().app_helper.get_attributes("gateway_attr_gateway")
-    stub: user_pb2_grpc.UserStub = pait_context.get().app_helper.get_attributes("gateway_attr_User")
+    gateway = pait_context.get().app_helper.get_attributes("gateway_attr_user_by_option_gateway")
     request_dict: dict = request_pydantic_model.dict()
     request_dict["token"] = token
-    request_msg: user_pb2.LogoutUserRequest = gateway.get_msg_from_dict(user_p2p.LogoutUserRequest, request_dict)
-    grpc_msg: Empty = stub.logout_user(request_msg, metadata=[("req_id", req_id)])
+    request_msg: user_pb2.LogoutUserRequest = gateway.get_msg_from_dict(user_pb2.LogoutUserRequest, request_dict)
+    grpc_msg: Empty = gateway.User_stub.logout_user(request_msg, metadata=[("req_id", req_id)])
     return gateway.make_response(gateway.msg_to_dict(grpc_msg))
 
 
 async def async_login_user_route(request_pydantic_model: user_p2p.LoginUserRequest) -> Any:
-    gateway = pait_context.get().app_helper.get_attributes("gateway_attr_gateway")
-    stub: user_pb2_grpc.UserStub = pait_context.get().app_helper.get_attributes("gateway_attr_User")
+    gateway = pait_context.get().app_helper.get_attributes("gateway_attr_user_by_option_gateway")
     request_msg: user_pb2.LoginUserRequest = gateway.get_msg_from_dict(
-        user_p2p.LoginUserRequest, request_pydantic_model.dict()
+        user_pb2.LoginUserRequest, request_pydantic_model.dict()
     )
     loop: asyncio.AbstractEventLoop = asyncio.get_event_loop()
-    if loop != stub.login_user._loop:  # type: ignore
+    if loop != gateway.User_stub.login_user._loop:  # type: ignore
         raise RuntimeError(
             "Loop is not same, "
             "the grpc channel must be initialized after the event loop of the api server is initialized"
         )
     else:
-        grpc_msg: user_pb2.LoginUserResult = await stub.login_user(request_msg)
+        grpc_msg: user_pb2.LoginUserResult = await gateway.User_stub.login_user(request_msg)
     return gateway.make_response(gateway.msg_to_dict(grpc_msg))
 
 
 def login_user_route(request_pydantic_model: user_p2p.LoginUserRequest) -> Any:
-    gateway = pait_context.get().app_helper.get_attributes("gateway_attr_gateway")
-    stub: user_pb2_grpc.UserStub = pait_context.get().app_helper.get_attributes("gateway_attr_User")
+    gateway = pait_context.get().app_helper.get_attributes("gateway_attr_user_by_option_gateway")
     request_msg: user_pb2.LoginUserRequest = gateway.get_msg_from_dict(
-        user_p2p.LoginUserRequest, request_pydantic_model.dict()
+        user_pb2.LoginUserRequest, request_pydantic_model.dict()
     )
-    grpc_msg: user_pb2.LoginUserResult = stub.login_user(request_msg)
+    grpc_msg: user_pb2.LoginUserResult = gateway.User_stub.login_user(request_msg)
     return gateway.make_response(gateway.msg_to_dict(grpc_msg))
 
 
 async def async_create_user_route(request_pydantic_model: user_p2p.CreateUserRequest) -> Any:
-    gateway = pait_context.get().app_helper.get_attributes("gateway_attr_gateway")
-    stub: user_pb2_grpc.UserStub = pait_context.get().app_helper.get_attributes("gateway_attr_User")
+    gateway = pait_context.get().app_helper.get_attributes("gateway_attr_user_by_option_gateway")
     request_msg: user_pb2.CreateUserRequest = gateway.get_msg_from_dict(
-        user_p2p.CreateUserRequest, request_pydantic_model.dict()
+        user_pb2.CreateUserRequest, request_pydantic_model.dict()
     )
     loop: asyncio.AbstractEventLoop = asyncio.get_event_loop()
-    if loop != stub.create_user._loop:  # type: ignore
+    if loop != gateway.User_stub.create_user._loop:  # type: ignore
         raise RuntimeError(
             "Loop is not same, "
             "the grpc channel must be initialized after the event loop of the api server is initialized"
         )
     else:
-        grpc_msg: Empty = await stub.create_user(request_msg)
+        grpc_msg: Empty = await gateway.User_stub.create_user(request_msg)
     return gateway.make_response(gateway.msg_to_dict(grpc_msg))
 
 
 def create_user_route(request_pydantic_model: user_p2p.CreateUserRequest) -> Any:
-    gateway = pait_context.get().app_helper.get_attributes("gateway_attr_gateway")
-    stub: user_pb2_grpc.UserStub = pait_context.get().app_helper.get_attributes("gateway_attr_User")
+    gateway = pait_context.get().app_helper.get_attributes("gateway_attr_user_by_option_gateway")
     request_msg: user_pb2.CreateUserRequest = gateway.get_msg_from_dict(
-        user_p2p.CreateUserRequest, request_pydantic_model.dict()
+        user_pb2.CreateUserRequest, request_pydantic_model.dict()
     )
-    grpc_msg: Empty = stub.create_user(request_msg)
+    grpc_msg: Empty = gateway.User_stub.create_user(request_msg)
     return gateway.make_response(gateway.msg_to_dict(grpc_msg))
 
 
@@ -130,14 +131,21 @@ async def async_delete_user_route(
     token: str = Header.i(description="User Token"),
     req_id: str = Header.i(alias="X-Request-Id", default_factory=lambda: str(uuid4())),
 ) -> Any:
-    gateway = pait_context.get().app_helper.get_attributes("gateway_attr_gateway")
-    stub: user_pb2_grpc.UserStub = pait_context.get().app_helper.get_attributes("gateway_attr_User")
+    gateway = pait_context.get().app_helper.get_attributes("gateway_attr_user_by_option_gateway")
+    stub: user_pb2_grpc.UserStub = gateway.User_stub
+    loop: asyncio.AbstractEventLoop = asyncio.get_event_loop()
+    if loop != stub.delete_user._loop:  # type: ignore
+        raise RuntimeError(
+            "Loop is not same, "
+            "the grpc channel must be initialized after the event loop"
+            "of the api server is initialized"
+        )
     # check token
     result: user_pb2.GetUidByTokenResult = await stub.get_uid_by_token(user_pb2.GetUidByTokenRequest(token=token))
     if not result.uid:
         raise RuntimeError("Not found user by token:" + token)
     request_msg: user_pb2.DeleteUserRequest = gateway.get_msg_from_dict(
-        user_p2p.DeleteUserRequest, request_pydantic_model.dict()
+        user_pb2.DeleteUserRequest, request_pydantic_model.dict()
     )
     grpc_msg: Empty = await stub.delete_user(request_msg, metadata=[("req_id", req_id)])
     return gateway.make_response(gateway.msg_to_dict(grpc_msg))
@@ -148,23 +156,25 @@ def delete_user_route(
     token: str = Header.i(description="User Token"),
     req_id: str = Header.i(alias="X-Request-Id", default_factory=lambda: str(uuid4())),
 ) -> Any:
-    gateway = pait_context.get().app_helper.get_attributes("gateway_attr_gateway")
-    stub: user_pb2_grpc.UserStub = pait_context.get().app_helper.get_attributes("gateway_attr_User")
+    gateway = pait_context.get().app_helper.get_attributes("gateway_attr_user_by_option_gateway")
+    stub: user_pb2_grpc.UserStub = gateway.User_stub
     # check token
     result: user_pb2.GetUidByTokenResult = stub.get_uid_by_token(user_pb2.GetUidByTokenRequest(token=token))
     if not result.uid:
         raise RuntimeError("Not found user by token:" + token)
     request_msg: user_pb2.DeleteUserRequest = gateway.get_msg_from_dict(
-        user_p2p.DeleteUserRequest, request_pydantic_model.dict()
+        user_pb2.DeleteUserRequest, request_pydantic_model.dict()
     )
     grpc_msg: Empty = stub.delete_user(request_msg, metadata=[("req_id", req_id)])
     return gateway.make_response(gateway.msg_to_dict(grpc_msg))
 
 
 class StaticGrpcGatewayRoute(BaseStaticGrpcGatewayRoute):
+    User_stub: user_pb2_grpc.UserStub
+    stub_str_list: List[str] = ["User_stub"]
+
     def gen_route(self) -> None:
-        set_app_attribute(self.app, "gateway_attr_gateway", self)
-        set_app_attribute(self.app, "gateway_attr_User", user_pb2_grpc.UserStub(self.channel))
+        set_app_attribute(self.app, "gateway_attr_user_by_option_gateway", self)
         # The response model generated based on Protocol is important and needs to be put first
         response_model_list: List[Type[BaseResponseModel]] = self._pait.response_model_list or []
         logout_user_route_pait: Pait = self._pait.create_sub_pait(
@@ -172,59 +182,63 @@ class StaticGrpcGatewayRoute(BaseStaticGrpcGatewayRoute):
             group="user",
             append_tag=(
                 Tag("grpc-user", "grpc_user_service"),
-                Tag("grpc-user_by_option", ""),
+                Tag("grpc-user_by_option-User", ""),
+                self._grpc_tag,
             ),
             desc="",
             summary="User exit from the system",
             default_field_class=field.Body,
             response_model_list=[UserByOptionEmptyJsonResponseModel] + response_model_list,
         )
-        pait_async_logout_user_route = logout_user_route_pait()(async_logout_user_route)
-        pait_logout_user_route = logout_user_route_pait()(logout_user_route)
+        pait_async_logout_user_route = logout_user_route_pait(feature_code="0")(async_logout_user_route)
+        pait_logout_user_route = logout_user_route_pait(feature_code="0")(logout_user_route)
         login_user_route_pait: Pait = self._pait.create_sub_pait(
             name="",
             group="user",
             append_tag=(
                 Tag("grpc-user", "grpc_user_service"),
-                Tag("grpc-user_by_option", ""),
+                Tag("grpc-user_by_option-User", ""),
+                self._grpc_tag,
             ),
             desc="",
             summary="User login to system",
             default_field_class=field.Body,
             response_model_list=[UserByOptionLoginUserResultJsonResponseModel] + response_model_list,
         )
-        pait_async_login_user_route = login_user_route_pait()(async_login_user_route)
-        pait_login_user_route = login_user_route_pait()(login_user_route)
+        pait_async_login_user_route = login_user_route_pait(feature_code="0")(async_login_user_route)
+        pait_login_user_route = login_user_route_pait(feature_code="0")(login_user_route)
         create_user_route_pait: Pait = self._pait.create_sub_pait(
             name="",
             group="user",
             append_tag=(
                 Tag("grpc-user", "grpc_user_service"),
                 Tag("grpc-user-system", "grpc_user_service"),
-                Tag("grpc-user_by_option", ""),
+                Tag("grpc-user_by_option-User", ""),
+                self._grpc_tag,
             ),
             desc="",
             summary="Create users through the system",
             default_field_class=field.Body,
             response_model_list=[UserByOptionEmptyJsonResponseModel] + response_model_list,
         )
-        pait_async_create_user_route = create_user_route_pait()(async_create_user_route)
-        pait_create_user_route = create_user_route_pait()(create_user_route)
+        pait_async_create_user_route = create_user_route_pait(feature_code="0")(async_create_user_route)
+        pait_create_user_route = create_user_route_pait(feature_code="0")(create_user_route)
         delete_user_route_pait: Pait = self._pait.create_sub_pait(
             name="",
             group="user",
             append_tag=(
                 Tag("grpc-user", "grpc_user_service"),
                 Tag("grpc-user-system", "grpc_user_service"),
-                Tag("grpc-user_by_option", ""),
+                Tag("grpc-user_by_option-User", ""),
+                self._grpc_tag,
             ),
             desc="This interface performs a logical delete, not a physical delete",
             summary="",
             default_field_class=field.Body,
             response_model_list=[UserByOptionEmptyJsonResponseModel] + response_model_list,
         )
-        pait_async_delete_user_route = delete_user_route_pait()(async_delete_user_route)
-        pait_delete_user_route = delete_user_route_pait()(delete_user_route)
+        pait_async_delete_user_route = delete_user_route_pait(feature_code="0")(async_delete_user_route)
+        pait_delete_user_route = delete_user_route_pait(feature_code="0")(delete_user_route)
         add_multi_simple_route(
             self.app,
             SimpleRoute(
